Add CDS push functionality, unit tests
Adds CDS push functionality for pushing the exported strings to CDS.

Moves all necessary classes to a different target (`TXCliLib`) so that
both the `TXCli` and the unit test targets can import it correctly and
use the classes.

Updates documentation and renames `project-name` argument to a more
simplified `project`.

Adds a unit test for testing XLIFFParser functionality.

Creates a helper class to deal with verbose logging.

Adds extra argument for keeping the temporary folder if needed.