Add pull command
Adds the pull command that downloads any translations that the user
specifies and stores them to disk in a folder specified by the user.

In order for both push and pull commands to be able to share a number
of arguments, a `ParsableArguments` struct had to be created that's now
used by both `ParsableCommand` structs.

The `project` argument of the push command is now required and the
logger logs the value of the purge flag for better readability.

The pull command downloads the translations from CDS and by using the
`fetchTranslations()` completion block, reads the downloaded
translations and serializes them into a JSON file, while also printing
out helpful instructions for the developer in order for the file to be
added in the app bundle.