Allow direct XLIFF parsing
The `--project` argument of the push command now also supports parsing
directly from an `.xliff` file, allowing anyone with access to the
generated `.xliff` file to push strings to CDS.

This can be helpful if access to the source code of the project is not
available or if the step that performs the automatic XLIFF generation
from the Xcode project has to be skipped for some reason.

If instead of the `.xliff` file, the path to the `.xcodeproj` container
is provided, the logic will behave exactly as before.