Parse strings with special characters correctly
When a source string contained a special character (e.g. a quote), the
parsing was dropping all characters before that special character,
losing part of the string.

In this commit, we ensure that we append to the active translation unit
instead of overwriting it, which updates the currently parsed string
correctly.

A new unit test has been added to test against this particular case
using a number of different special characters.