Support xcworkspace on project parameter
The cli tool was not detecting the `.xcworkspace` projects, leading to
cases where users had to provide the `.xcodeproj` file instead but that
does not always work, as the underlying `xcodebuild` tool will try to
build the project, missing all dependencies that may be needed.

This commit makes it possible for the cli tool to detect when users
provide a `.xcworkspace` instead of a `.xcodeproj` and use the correct
argument for the `xcodebuild` tool.

The help string has been also updated to reflect that change and the
logic now fails earlier if users provide a file that is neither a
`.xliff`, `.xcodeproj` or a `.xcworkspace` one.