Introduce status filter option in pull command
Introduces the `--with-status-only` option that can be provided
alongside the existing `--with-tags-only` option and allows users to
have better control over which strings will be downloaded from CDS.

The `--with-status-only` option accepts a single string that represents
the status that the pulled strings must have in order to be downloaded.

Points to version 1.0.4 of Transifex SDK so that the CLI tool can have
access to the status filter.

This commit also improves the documentation of the `hashKeys` option, so
that it's clear what the `true` and `false` values do exactly.