Introduce excluded files option
Introduces the `--excluded-files` option in the `push` command so that
users can offer an optional list of filenames that the internal logic
must exclude when processing the intermediate XLIFF file.

Developers can offer a comma separated list of filenames that must be
filtered out by the CLI before the strings are pushed to CDS.

Example:

```
txios-cli push ... --excluded-files ExcludedFile1.strings,ExcludedFile2.strings
```

Note: If a string is included both in the excluded filenames and is
also found in non-excluded filenames, it will be included in the push
data.

Internally, the logic appends this optional list to the list of
unsupported files that is already being filtered out.

The filtered strings are logged in case user has enabled the
verbose logging option.