Fix excluded filenames issue
The logic responsible for excluding the filename list provided by the
developer via the respective option of the push command was not
including the `UNSUPPORTED_FILES`, thus not excluding files that are not
supported by the SDK like `InfoPlist.strings` or `Root.string`.

This commit fixes this issue by passing the correct array to the logic.