Address language tag discrepancy
iOS accesses the language tags using a hyphen (e.g. `en-GB`), whilst
Transifex uses an underscore for the delimiter (e.g. `en_GB`).

This discrepancy can cause confusion to developers which expect
the Transifex format and will try to pass that in the `--source-locale`
option of the `push` command, causing the push logic to fail to export
the localizations from the project.

This commit addresses the issue by normalizing the `sourceLocale`
argument in the `LocalizationExporter` constructor so that the
underscore character (`_`) is replaced with a hypen (`-`) so that the
`xcodebuild` command will be able to export the localizations
correctly.