Add base SDK option for push command
When exporting localizations via the `push` command, the default SDK
that was being used was the `macosx` one, resulting to errors when
the project included Swift packages using iOS frameworks.

In order to address this and provide more control to the developer, a
new option was introduced in the `push` command: `--base-sdk`.

The value of this option is passed as the `-sdk` option of the
`xcodebuild` process that is called to export the project's
localizations (via the `-exportLocalizations` option).

If no value is provided to this option, then the value of the `-sdk`
option is also not specified in the `xcodebuild` process and the default
SDK is picked by the system.