Resolve --keep-translations argument issue
The `--keep-translations` argument of the `push` command was a boolean
flag that had a default value of `true`. This meant that if user did
not supply the `--keep-translations` argument to their `push` command
the value would be `true` and if they did, the value would also be
`true`, making it impossible for the `--keep-translations` flag to be
false.

In order to fix the issue, the argument has been replaced with the
inverse flag: `--delete-translations`. Its default value is `false` so
it behaves exactly as the `--keep-translations` before, if the argument
was not supplied (`keep_translations` is going to be set to `true`).

The difference is that now if this new argument is supplied, then the
underlying value of the `keep_translations` metadata will be `false`.