Full String Catalog support
With the introduction of the String Catalogs file format in Xcode 15,
developers have now the option to create localizations that vary by
device and/or by plural. On top of that, multiple tokens can be added in
a localized phrase (also known as substitutions) that can also support
the aforementioned variations.

In order to support all those new formats as well as the old
substitution rules found in the Strings Dictionary format
(`.stringsdict`), the advanced rules from the CLI tool are formatted as
an intermediate XML structure and pushed to CDS.

This commit adds this functionality so that complex variation rules can
be supported and displayed in the Transifex web interface.

More unit tests have been added to test those cases as well as edge-case
scenarios and existing unit tests have been improved.